Add generate button to generate graph from input, clean up util
